moved the baseOptions function to Giveaways class
