+++ src/classes/main.js
@@ -4,25 +4,7 @@ let win;
 const schema = require('../models/giveawayschema');
 const { MessageButton } = require('discord-buttons');
 const giveaways = require('./giveaways');
-const merge = require('deepmerge');
-const defaultManagerOptions = {
-    dmWinner: true,
-    giveaway: '🎉🎉 **GIVEAWAY MOMENT** 🎉🎉',
-    giveawayDescription: '🎁 Prize: **${prize}**\n🎊 Hosted by: ${hostedBy}\n⏲️ Winner(s): \`{winners}\`\n\nRequirements: {requirements}',
-    endedGiveawayDescription : '🎁 Prize: **{prize}**\n🎊 Hosted by: ${hostedBy}\n⏲️ Winner(s): {winners}',
-    giveawayFooterImage: 'https://cdn.discordapp.com/emojis/843076397345144863.png',
-    winMessage: '{winners} you won {prize} Congratulations! Hosted by {hostedBy}',
-	rerolledMessage: 'Rerolled! {winner} is the new winner of the giveaway!', // only {winner} placeholder
-    toParticipate: '**Click the Enter button to enter the giveaway!**',
-	newParticipant: 'You have successfully entered for this giveaway', // no placeholders | ephemeral
-	alreadyParticipated: 'you already entered this giveaway!', // no placeholders | ephemeral
-	nonoParticipants: 'There are not enough people in the giveaway!', // no placeholders
-	nonoRole: 'You do not have the required role(s)\n{requiredRoles}\n for the giveaway!', // only {requiredRoles} | ephemeral
-    dmMessage: 'You have won a giveaway in **{guildName}**!\nPrize: [{prize}]({giveawayURL})',
-    noWinner: 'Not enough people participated in this giveaway.', // no {winner} placerholder
-    alreadyEnded: 'The giveaway has already ended!', // no {winner} placeholder
-    dropWin: '{winner} Won The Drop!!' // only {winner} placeholder
-}
+
 class main {
 /**
 	*
@@ -34,10 +16,6 @@ class main {
 			useNewUrlParser: true,
 			useUnifiedTopology: true,
 		});
-	}
-	static async baseOptions(client, options) {
-		this.client = client
-		client.giveawayMessages = merge(defaultManagerOptions, options);			
 	}
 	static async buttonclick(client, button) {
 		if (!client) throw new Error('NuggiesError: client not provided');
